API/Docker_requirements.txt
pillow
aiocoap
requests
Flask
paho-mqtt